+++ src/examples/zkapps/dex/run.ts
@@ -9,7 +9,6 @@ let Local = Mina.LocalBlockchain({
   enforceTransactionLimits: false,
 });
 Mina.setActiveInstance(Local);
-let accountFee = Mina.accountCreationFee();
 let [{ privateKey: feePayerKey, publicKey: feePayerAddress }] =
   Local.testAccounts;
 let tx, balances, oldBalances;
@@ -77,6 +76,7 @@ async function main({ withVesting }: { withVesting: boolean }) {
 
   console.log('deploy & init token contracts...');
   tx = await Mina.transaction(feePayerAddress, () => {
+    const accountFee = Mina.getNetworkConstants().accountCreationFee;
     // pay fees for creating 2 token contract accounts, and fund them so each can create 2 accounts themselves
     let feePayerUpdate = AccountUpdate.fundNewAccount(feePayerAddress, 2);
     feePayerUpdate.send({ to: addresses.tokenX, amount: accountFee.mul(2) });
@@ -110,7 +110,10 @@ async function main({ withVesting }: { withVesting: boolean }) {
 
   console.log('transfer tokens to user');
   tx = await Mina.transaction(
-    { sender: feePayerAddress, fee: accountFee.mul(1) },
+    {
+      sender: feePayerAddress,
+      fee: Mina.getNetworkConstants().accountCreationFee.mul(1),
+    },
     () => {
       let feePayer = AccountUpdate.fundNewAccount(feePayerAddress, 4);
       feePayer.send({ to: addresses.user, amount: 20e9 }); // give users MINA to pay fees
@@ -133,7 +136,10 @@ async function main({ withVesting }: { withVesting: boolean }) {
   // supply the initial liquidity where the token ratio can be arbitrary
   console.log('supply liquidity -- base');
   tx = await Mina.transaction(
-    { sender: feePayerAddress, fee: accountFee },
+    {
+      sender: feePayerAddress,
+      fee: Mina.getNetworkConstants().accountCreationFee,
+    },
     () => {
       AccountUpdate.fundNewAccount(feePayerAddress);
       dex.supplyLiquidityBase(UInt64.from(10_000), UInt64.from(10_000));
@@ -437,7 +443,7 @@ async function main({ withVesting }: { withVesting: boolean }) {
   );
   tx = await Mina.transaction(addresses.user2, () => {
     AccountUpdate.createSigned(addresses.user2).balance.subInPlace(
-      accountFee.mul(2)
+      Mina.getNetworkConstants().accountCreationFee.mul(2)
     );
     dex.redeemLiquidity(UInt64.from(USER_DL));
     dex.redeemLiquidity(UInt64.from(USER_DL));
@@ -451,7 +457,7 @@ async function main({ withVesting }: { withVesting: boolean }) {
   console.log('user2 redeem liquidity');
   tx = await Mina.transaction(addresses.user2, () => {
     AccountUpdate.createSigned(addresses.user2).balance.subInPlace(
-      accountFee.mul(2)
+      Mina.getNetworkConstants().accountCreationFee.mul(2)
     );
     dex.redeemLiquidity(UInt64.from(USER_DL));
   });