+++ src/lib/gadgets/gadgets.ts
@@ -24,6 +24,7 @@ import {
 import { Field } from '../core.js';
 import { ForeignField, Field3, Sum } from './foreign-field.js';
 import { divMod32, addMod32 } from './arithmetic.js';
+import { SHA256 } from './sha256.js';
 
 export { Gadgets };
 
@@ -796,6 +797,26 @@ const Gadgets = {
    * ```
    *    */
   addMod32,
+
+  /**
+   * Implementation of the [SHA256 hash function.](https://en.wikipedia.org/wiki/SHA-2) Hash function with 256bit output.
+   *
+   * Applies the SHA2-256 hash function to a list of byte-sized elements.
+   *
+   * The function accepts {@link Bytes} as the input message, which is a type that represents a static-length list of byte-sized field elements (range-checked using {@link Gadgets.rangeCheck8}).
+   * Alternatively, you can pass plain `number[]`, `bigint[]` or `Uint8Array` to perform a hash outside provable code.
+   *
+   * Produces an output of {@link Bytes} that conforms to the chosen bit length.
+   *
+   * @param data - {@link Bytes} representing the message to hash.
+   *
+   * ```ts
+   * let preimage = Bytes.fromString("hello world");
+   * let digest = Gadgets.SHA256.hash(preimage);
+   * ```
+   *
+   */
+  SHA256: SHA256,
 };
 
 export namespace Gadgets {