+++ src/lib/mina.ts
@@ -51,6 +51,7 @@ export {
   getAccount,
   hasAccount,
   getBalance,
+  getNetworkConstants,
   getNetworkState,
   accountCreationFee,
   sendTransaction,
@@ -64,7 +65,9 @@ export {
   getProofsEnabled,
   // for internal testing only
   filterGroups,
+  type NetworkConstants
 };
+
 interface TransactionId {
   isSuccess: boolean;
   wait(options?: { maxAttempts?: number; interval?: number }): Promise<void>;
@@ -161,6 +164,15 @@ type DeprecatedFeePayerSpec =
 type ActionStates = {
   fromActionState?: Field;
   endActionState?: Field;
+};
+
+type NetworkConstants = {
+  genesisTimestamp: UInt64;
+  /**
+   * Duration of 1 slot in milliseconds
+   */
+  slotTime: UInt64;
+  accountCreationFee: UInt64;
 };
 
 function reportGetAccountError(publicKey: string, tokenId: string) {
@@ -340,14 +352,10 @@ interface Mina {
   hasAccount(publicKey: PublicKey, tokenId?: Field): boolean;
   getAccount(publicKey: PublicKey, tokenId?: Field): Account;
   getNetworkState(): NetworkValue;
-  getNetworkConstants(): {
-    genesisTimestamp: UInt64;
-    /**
-     * Duration of 1 slot in millisecondw
-     */
-    slotTime: UInt64;
-    accountCreationFee: UInt64;
-  };
+  getNetworkConstants(): NetworkConstants;
+  /**
+   * @deprecated use {@link getNetworkConstants}
+   */
   accountCreationFee(): UInt64;
   sendTransaction(transaction: Transaction): Promise<TransactionId>;
   fetchEvents: (
@@ -369,21 +377,23 @@ interface Mina {
 }
 
 const defaultAccountCreationFee = 1_000_000_000;
+const defaultNetworkConstants: NetworkConstants = {
+  genesisTimestamp: UInt64.from(0),
+  slotTime: UInt64.from(3 * 60 * 1000),
+  accountCreationFee: UInt64.from(defaultAccountCreationFee),
+};
 
 /**
  * A mock Mina blockchain running locally and useful for testing.
  */
 function LocalBlockchain({
-  accountCreationFee = defaultAccountCreationFee as string | number,
   proofsEnabled = true,
   enforceTransactionLimits = true,
 } = {}) {
   const slotTime = 3 * 60 * 1000;
   const startTime = Date.now();
   const genesisTimestamp = UInt64.from(startTime);
-
   const ledger = Ledger.create();
-
   let networkState = defaultNetworkState();
 
   function addAccount(publicKey: PublicKey, balance: string) {
@@ -412,12 +422,14 @@ function LocalBlockchain({
 
   return {
     proofsEnabled,
-    accountCreationFee: () => UInt64.from(accountCreationFee),
+    /**
+   * @deprecated use {@link Mina.getNetworkConstants}
+   */
+    accountCreationFee: () => defaultNetworkConstants.accountCreationFee,
     getNetworkConstants() {
       return {
+        ...defaultNetworkConstants,
         genesisTimestamp,
-        accountCreationFee: UInt64.from(accountCreationFee),
-        slotTime: UInt64.from(slotTime),
       };
     },
     currentSlot() {
@@ -490,7 +502,7 @@ function LocalBlockchain({
       try {
         ledger.applyJsonTransaction(
           JSON.stringify(zkappCommandJson),
-          String(accountCreationFee),
+          defaultNetworkConstants.accountCreationFee.toString(),
           JSON.stringify(networkState)
         );
       } catch (err: any) {
@@ -499,7 +511,7 @@ function LocalBlockchain({
           // TODO: label updates, and try to give precise explanations about what went wrong
           let errors = JSON.parse(err.message);
           err.message = invalidTransactionError(txn.transaction, errors, {
-            accountCreationFee,
+            accountCreationFee: defaultNetworkConstants.accountCreationFee.toString(),
           });
         } finally {
           throw err;
@@ -601,7 +613,7 @@ function LocalBlockchain({
     applyJsonTransaction(json: string) {
       return ledger.applyJsonTransaction(
         json,
-        String(accountCreationFee),
+        defaultNetworkConstants.accountCreationFee.toString(),
         JSON.stringify(networkState)
       );
     },
@@ -691,7 +703,6 @@ function Network(
       }
     | string
 ): Mina {
-  let accountCreationFee = UInt64.from(defaultAccountCreationFee);
   let minaGraphqlEndpoint: string;
   let archiveEndpoint: string;
   let lightnetAccountManagerEndpoint: string;
@@ -737,22 +748,30 @@ function Network(
     );
   }
 
-  // copied from mina/genesis_ledgers/berkeley.json
-  // TODO fetch from graphql instead of hardcoding
-  const genesisTimestampString = '2023-02-23T20:00:01Z';
-  const genesisTimestamp = UInt64.from(
-    Date.parse(genesisTimestampString.slice(0, -1) + '+00:00')
-  );
-  // TODO also fetch from graphql
-  const slotTime = UInt64.from(3 * 60 * 1000);
   return {
-    accountCreationFee: () => accountCreationFee,
+    /**
+   * @deprecated use {@link Mina.getNetworkConstants}
+   */
+    accountCreationFee: () => defaultNetworkConstants.accountCreationFee,
     getNetworkConstants() {
-      return {
-        genesisTimestamp,
-        slotTime,
-        accountCreationFee,
-      };
+      if (currentTransaction()?.fetchMode === 'test') {
+        Fetch.markNetworkToBeFetched(minaGraphqlEndpoint);
+        const genesisConstants =
+          Fetch.getCachedGenesisConstants(minaGraphqlEndpoint);
+        return genesisConstants !== undefined
+          ? genesisToNetworkConstants(genesisConstants)
+          : defaultNetworkConstants;
+      }
+      if (
+        !currentTransaction.has() ||
+        currentTransaction.get().fetchMode === 'cached'
+      ) {
+        const genesisConstants =
+          Fetch.getCachedGenesisConstants(minaGraphqlEndpoint);
+        if (genesisConstants !== undefined)
+          return genesisToNetworkConstants(genesisConstants);
+      }
+      return defaultNetworkConstants;
     },
     currentSlot() {
       throw Error(
@@ -814,7 +833,7 @@ function Network(
         if (network !== undefined) return network;
       }
       throw Error(
-        `getNetworkState: Could not fetch network state from graphql endpoint ${minaGraphqlEndpoint}`
+        `getNetworkState: Could not fetch network state from graphql endpoint ${minaGraphqlEndpoint} outside of a transaction.`
       );
     },
     async sendTransaction(txn: Transaction) {
@@ -994,9 +1013,12 @@ function BerkeleyQANet(graphqlEndpoint: string) {
 }
 
 let activeInstance: Mina = {
-  accountCreationFee: () => UInt64.from(defaultAccountCreationFee),
+  /**
+   * @deprecated use {@link Mina.getNetworkConstants}
+   */
+  accountCreationFee: () => defaultNetworkConstants.accountCreationFee,
   getNetworkConstants() {
-    throw new Error('must call Mina.setActiveInstance first');
+    return defaultNetworkConstants;
   },
   currentSlot: () => {
     throw new Error('must call Mina.setActiveInstance first');
@@ -1178,6 +1200,13 @@ function hasAccount(publicKey: PublicKey, tokenId?: Field): boolean {
 }
 
 /**
+ * @return Data associated with the current Mina network constants.
+ */
+function getNetworkConstants() {
+  return activeInstance.getNetworkConstants();
+}
+
+/**
  * @return Data associated with the current state of the Mina network.
  */
 function getNetworkState() {
@@ -1193,6 +1222,7 @@ function getBalance(publicKey: PublicKey, tokenId?: Field) {
 
 /**
  * Returns the default account creation fee.
+ * @deprecated use {@link Mina.getNetworkConstants}
  */
 function accountCreationFee() {
   return activeInstance.accountCreationFee();
@@ -1591,4 +1621,16 @@ async function faucet(pub: PublicKey, network: string = 'berkeley-qanet') {
     );
   }
   await waitForFunding(address);
+}
+
+function genesisToNetworkConstants(
+  genesisConstants: Fetch.GenesisConstants
+): NetworkConstants {
+  return {
+    genesisTimestamp: UInt64.from(
+      Date.parse(genesisConstants.genesisTimestamp)
+    ),
+    slotTime: UInt64.from(genesisConstants.slotDuration),
+    accountCreationFee: UInt64.from(genesisConstants.accountCreationFee),
+  };
 }